HoiOriClassifier/requirements.txt
numpy
tqdm
Pillow
torch
torchvision
transformers
PyQt5
scipy